Moved assembly classes to parsers
