+++ src/common/index.ts
@@ -1,6 +1,3 @@
-export * from './assembly-block';
-export * from './assembly-file-reference';
-export * from './assembly-symbol';
 export * from './convert';
 export * from './guards';
 export * from './language';
