+++ src/parsers/assembly-block.ts
