+++ src/parsers/assembly-file-reference.ts
