+++ src/parsers/index.ts
@@ -1,4 +1,7 @@
+export * from './assembly-block';
 export * from './assembly-document';
+export * from './assembly-file-reference';
 export * from './assembly-line';
+export * from './assembly-symbol';
 export * from './docs';
 export * from './line-parser';
